Extract chat personality helpers
