+++ service-worker.js
@@ -73,6 +73,7 @@ const APP_SHELL = [
   '/js/chat.js',
   '/js/chat-attestation.js',
   '/js/chat-icons.js',
+  '/js/chat-personalities.js',
   '/js/chat-continuation.js',
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
